+++ test/operators.jl
@@ -83,6 +83,14 @@ import Base.<
 
 @test isless('a','b')
 
+@testset "isgreater" begin
+    @test !isgreater(missing, 1)
+    @test isgreater(5, 1)
+    @test !isgreater(1, 5)
+    @test isgreater(1, missing)
+    @test isgreater(1, NaN)
+end
+
 @testset "vectorized comparisons between numbers" begin
     @test 1 .!= 2
     @test 1 .== 1